backend/app/pyproject.toml
[tool.poetry]
name = "Storious"
version = "0.1.0"
description = ""
authors = ["Valerii Platonov <valeri.platonov@gmail.com>"]
readme = "README.md"
package-mode = false
#packages = [{include = "Storious"}]

[tool.poetry.dependencies]
python = "^3.10"
typing_extensions = "^4.12.2"
typing_inspect = "^0.9.0"
fastapi-utils = "^0.7.0"
python-multipart = "^0.0.9"
Authlib = "^1.3.1"
requests = "^2.32.3"
itsdangerous = "^2.1.1"
dacite = "^1.6.0"
pyjwt = "^1.7.1"
httpx = "^0.27.0"

shortuuid = "^1.0.8"
croniter = "^1.3.4"
dnspython = "^2.4.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
pydantic-argparse = "^0.9.0"

azure-core = "^1.29.1"
azure-keyvault-keys = "^4.8.0"
azure-identity = "^1.14.1"
azure-mgmt-appcontainers = "^3.0.0"
pymssql = "^2.3.0"
pyodbc = "^5.1.0"

[tool.poetry.group.test]  # This part can be left out
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-mock = "*"
pytest-aio = "^1.9.0"
pytest-env = "^1.1.3"

#[build-system]
#requires = ["poetry-core"]
#build-backend = "poetry.core.masonry.api"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
